Minor fixes and rearranging
Rearrange the file/folder structure of the website. The Tools section should be much more organized now, almost as if you can make a folder and use it as the workspace for an HTML webpage, along with its own virtual styles, scripts, and assets folders.

Also fixed the QR Code Generator link that was linking to the alt caser webpage.